+++ a10_octavia/controller/worker/flows/vthunder_flows.py
@@ -231,7 +231,7 @@ class VThunderFlows(object):
         vthunder_for_amphora_subflow.add(database_tasks.CreateAmphoraInDB(
             name=sf_name + '-' + constants.CREATE_AMPHORA_INDB,
             provides=constants.AMPHORA_ID))
-        vthunder_for_amphora_subflow.add(compute_tasks.ValidateComputeForProject(
+        vthunder_for_amphora_subflow.add(a10_database_tasks.ValidateComputeForProject(
             name=sf_name + '-' + a10constants.VALIDATE_COMPUTE_FOR_PROJECT,
             requires=constants.LOADBALANCER,
             inject={"role": role},
